+++ win/win32/mswproc.c
@@ -3060,7 +3060,7 @@ mswin_status_update(int idx, genericptr_t ptr, int chg, int percent, int color, 
     char *text = (char *) ptr;
     MSNHMsgUpdateStatus update_cmd_data;
     int ocolor, ochar;
-    unsigned ospecial;
+    unsigned long ospecial;
 
     logDebug("mswin_status_update(%d, %p, %d, %d, %x, %p)\n", idx, ptr, chg, percent, color, condmasks);
